+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.0.9] - 2026-Apr-13
 - NFTables address-list/set conversion function
 - NFT config output table builder
+- ratelimit rule splitter generator (1 nfty = 2 nftables)
+- rulebuilder that combines all lower-level rule tools to write fully-functional NFT-syntax rule
 
 ## [0.0.8] - 2026-Apr-13
 - basic structure for nfty → NFTables config generation